% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metaeco.R
\name{Metaeco_fun}
\alias{Metaeco_fun}
\title{Metaeco function}
\usage{
Metaeco_fun(year, month, day)
}
\arguments{
\item{year}{Enter the year in an int format, eg. 1989}

\item{month}{Enter the month in an int format, eg. 8}

\item{day}{Enter the day in an int format, eg. 25}
}
\description{
This function allows you to calculate the day based on date.
The formula used in the function is kim Larsson calculation formula.
}
\examples{
Metaeco_fun(2022, 7, 24)

}
\keyword{Metaeco}
